analysis/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
supabase>=0.0.1
python-dotenv>=0.19.0 